+++ app/src/main/java/com/example/myfinance/feature/domain/model/Transaction.kt
@@ -1,9 +1,10 @@
 package com.example.myfinance.feature.domain.model
 
+
 data class Transaction(
     val id: Int,
     val category: String,
-    val amount: Int,
+    val amount: Double,
     val date: String,
     val emoji: String = "",
     val comment: String = ""
